backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
pymongo==4.10.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# RAG dependencies (from retriever/)
langchain==0.3.12
langchain-google-genai==2.0.6
langchain-community==0.3.12
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
